requirements.txt
pandas==2.2.3
requests-html==0.10.0
beautifulsoup4==4.12.3
selenium==4.25.0
undetected-chromedriver==3.5.5
python-telegram-bot==21.5
geopy==2.4.1
python-dotenv==1.1.1